+++ src/utils/vscode/getInitialHTML.ts
@@ -124,8 +124,6 @@ export default function getInitialHTML(
             <button class='run-watermelon btn btn-primary' type='button'>Get Code Context</button>
           </div>
           <div id="jiraHolder"></div>
-          <div id="dailySummary"></div>
-          You will find all your issues and ToDos here. 
         </div>
      </body>
      <footer>